*updated combat - added wipe out mechanics

+++ Assets/Scripts/BalanceOfPower/SeeksSuperiority.cs
@@ -313,7 +313,7 @@ public static class SeeksSuperiority {
 								int chance = UnityEngine.Random.Range (0, 100);
 								int value = 1 * stabilityModifier;
 								int threshold = KingdomManager.Instance.GetReducedInvasionValueThreshHold (50f, overPopulationReduction);
-								int totalValue = (kr.targetKingdomInvasionValue - threshold) * value;
+								int totalValue = ((int)kr.targetKingdomInvasionValue - threshold) * value;
 								if(chance < value){
 									//if there is anyone whose Invasion Value is 1 or above, prepare for war against the one with the highest Invasion Value
 									Warfare warfare = new Warfare (kingdom, targetKingdom);
